Refactor About extension for multiple repos and host mention
The new functionality will allow multiple repos to be referenced,
however authors will no longer be automatically loaded as the list may
contain authors from separate repos and may grow excessively large.

There is also a new host option which will be resolved to a user
mention and included in the about message.